tools: move elektrad into /tools folder
