+++ src/tools/elektrad/build.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+cd @CMAKE_INSTALL_PREFIX@/@install_directory@
+
+cd elektrad
+go install .
+cd ..
