+++ src/tools/elektrad/copy_handler.go
@@ -1,0 +1,1 @@
+package main
