+++ src/tools/elektrad/go.mod
@@ -1,0 +1,9 @@
+module github.com/ElektraInitiative/libelektra/elektrad
+
+go 1.13
+
+require (
+	github.com/google/uuid v1.1.1
+	github.com/gorilla/mux v1.7.3
+	go.libelektra.org v0.0.0-20200214215340-d51fc6073e5d
+)
