+++ src/tools/elektrad/run-elektrad.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+cd @CMAKE_INSTALL_PREFIX@/@install_directory@
+
+./elektrad
