+++ src/tools/elektrad/temp-elektra.pc.in
@@ -1,0 +1,10 @@
+prefix=@CMAKE_BINARY_DIR@
+libdir=${prefix}/lib
+srcinclude=${prefix}/src/include
+buildinclude=@CMAKE_SOURCE_DIR@/src/include
+
+Name: @PACKAGE_NAME@
+Description: @PACKAGE_DESCRIPTION@
+Version: @KDB_VERSION@
+Cflags: -I${srcinclude} -I${buildinclude}
+Libs: -L${libdir} -lelektra-kdb -lelektra-core
