+++ src/tools/elektrad/test.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+cd @CMAKE_CURRENT_SOURCE_DIR@
+
+echo " -~- Testing API elektra-web > elektrad -~- "
+cd elektrad
+go test ./...
+cd ..
