--- src/tools/web/run-elektrad.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-cd @CMAKE_INSTALL_PREFIX@/@install_directory@
-
-./elektrad
